backend/pyproject.toml
[project]
name = "ai-resume-analyzer"
version = "1.0.0"
description = "AI Resume Analyzer - Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.44.0",
    "python-multipart>=0.0.26",
    "pdfplumber>=0.11.0",
    "python-docx>=1.2.0",
    "sentence-transformers>=5.4.0",
    "torch>=2.11.0",
]
